+++ ChangeLog.md
@@ -1,5 +1,5 @@
 #### v.4.0.6 (TBD)
-
+* Update to DICOM Standard 2020b.
 
 #### v.4.0.5 (5/18/2020)
 * Bug fix: DicomTags of ValueRepresentation LT have not been validated.